+++ proto/client-api/CHANGELOG.md
@@ -17,6 +17,7 @@ All user visible changes to this project will be documented in this file. This p
 
 - `PeerConnectionError` type ([#151]).
 - `IceCandidateError` type ([#151]).
+- `Codec`, `ScalabilityMode`, `SvcSetting` and `EncodingParameters`.
 
 [#151]: /../../pull/151